fix(#392): Cannot fetch bodies for multiple entities from multiple collections
When using multiple globally unique identifiers in single query to fetch entities where each entity is from different collection AND entity body is wanted, like so:
```
query(
  filterBy(
    attributeInSet("url", "/en/macbook-pro-13-2022", "/en/cell-phones"),
    entityLocaleEquals("en")
  ),
  require(
    entityFetch(
      attributeContent("code")
    )
  )
)
```
evitaDB returns error
```
Entity `Product` with primary key `63049` cannot be enriched because it has been already removed.
```
If both URLs would target the same collection, both entities would be returned with their bodies as expected.
