+++ evita_engine/src/main/java/io/evitadb/core/query/QueryPlan.java
@@ -36,7 +36,7 @@ import io.evitadb.api.requestResponse.data.structure.BinaryEntity;
 import io.evitadb.api.requestResponse.data.structure.EntityReference;
 import io.evitadb.api.requestResponse.extraResult.QueryTelemetry.QueryPhase;
 import io.evitadb.core.query.algebra.Formula;
-import io.evitadb.core.query.algebra.prefetch.SelectionFormula.PrefetchFormulaVisitor;
+import io.evitadb.core.query.algebra.prefetch.PrefetchFormulaVisitor;
 import io.evitadb.core.query.extraResult.ExtraResultProducer;
 import io.evitadb.core.query.sort.ConditionalSorter;
 import io.evitadb.core.query.sort.Sorter;
