+++ evita_engine/src/main/java/io/evitadb/core/query/QueryPlanBuilder.java
@@ -26,7 +26,7 @@ package io.evitadb.core.query;
 import io.evitadb.api.query.require.EntityContentRequire;
 import io.evitadb.core.query.algebra.Formula;
 import io.evitadb.core.query.algebra.base.EmptyFormula;
-import io.evitadb.core.query.algebra.prefetch.SelectionFormula.PrefetchFormulaVisitor;
+import io.evitadb.core.query.algebra.prefetch.PrefetchFormulaVisitor;
 import io.evitadb.core.query.extraResult.ExtraResultProducer;
 import io.evitadb.core.query.indexSelection.TargetIndexes;
 import io.evitadb.core.query.sort.NoSorter;
