+++ evita_engine/src/main/java/io/evitadb/core/query/QueryPlanner.java
@@ -36,8 +36,8 @@ import io.evitadb.core.query.algebra.AbstractFormula;
 import io.evitadb.core.query.algebra.Formula;
 import io.evitadb.core.query.algebra.base.NotFormula;
 import io.evitadb.core.query.algebra.debug.CacheableVariantsGeneratingVisitor;
+import io.evitadb.core.query.algebra.prefetch.PrefetchFormulaVisitor;
 import io.evitadb.core.query.algebra.prefetch.SelectionFormula;
-import io.evitadb.core.query.algebra.prefetch.SelectionFormula.PrefetchFormulaVisitor;
 import io.evitadb.core.query.extraResult.CacheDisabledExtraResultAccessor;
 import io.evitadb.core.query.extraResult.CacheTranslatingExtraResultAccessor;
 import io.evitadb.core.query.extraResult.CacheableExtraResultProducer;
